Move `ImmuTableRow` to `ImmuTable` package
