+++ Modules/Sources/ImmuTable/ImmuTableAction.swift
@@ -1,0 +1,1 @@
+public typealias ImmuTableAction = (ImmuTableRow) -> Void
