+++ WordPress/Classes/Utility/ImmuTable/ImmuTableViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressShared
