+++ WordPress/Classes/ViewRelated/Aztec/ViewControllers/AztecAttachmentViewController.swift
@@ -1,7 +1,8 @@
+import Aztec
 import Foundation
+import ImmuTable
 import UIKit
 import WordPressShared
-import Aztec
 
 class AztecAttachmentViewController: UITableViewController {
