+++ WordPress/Classes/ViewRelated/Aztec/ViewControllers/LinkSettingsViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 
 struct LinkSettings {
