+++ WordPress/Classes/ViewRelated/Jetpack/Jetpack Settings/JetpackConnectionViewController.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 
 @objc
 protocol JetpackConnectionDelegate {
