+++ WordPress/Classes/ViewRelated/Me/My Profile/MyProfileViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressShared
