+++ WordPress/Classes/ViewRelated/Plans/Controllers/PlanListViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressShared
