+++ WordPress/Classes/ViewRelated/Plans/ViewModels/PlanDetailViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 import WordPressShared
 
 struct PlanDetailViewModel {
