+++ WordPress/Classes/ViewRelated/Plans/ViewModels/PlanListViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 import WordPressShared
 import WordPressUI
