+++ WordPress/Classes/ViewRelated/Plugins/ViewModels/PluginViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 import WordPressFlux
 
 class PluginViewModel: Observable {
