+++ WordPress/Classes/ViewRelated/Stats/Insights/Insights Management/InsightsManagementViewController.swift
@@ -1,5 +1,6 @@
-import UIKit
 import Gridicons
+import ImmuTable
+import UIKit
 
 // This exists in addition to `SiteStatsInsightsDelegate` because `[StatSection]`
 // can't be represented in an Obj-C protocol.
