+++ WordPress/Classes/ViewRelated/Stats/Period Stats/HashableImmutableRow.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ImmuTable
 
 protocol HashableImmutableRow: ImmuTableRow, Hashable {}
