+++ WordPress/Classes/ViewRelated/Stats/Period Stats/SiteStatsPeriodTableViewController.swift
@@ -1,6 +1,7 @@
+import Combine
+import ImmuTable
 import UIKit
 import WordPressFlux
-import Combine
 
 @objc protocol SiteStatsPeriodDelegate {
     @objc optional func displayWebViewWithURL(_ url: URL)
