+++ WordPress/Classes/ViewRelated/Stats/Shared Views/Post Stats/PostStatsTableViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressFlux
