+++ WordPress/Classes/ViewRelated/Stats/Shared Views/Stats Detail/SiteStatsDetailTableViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressFlux
