+++ WordPress/Classes/ViewRelated/Stats/Shared Views/Stats Detail/SiteStatsInsightsDetailsTableViewController.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import UIKit
 import WordPressFlux
