+++ WordPress/Classes/ViewRelated/Tools/Time Zone/TimeZoneSelectorViewModel.swift
@@ -1,3 +1,4 @@
+import ImmuTable
 import WordPressFlux
 
 struct TimeZoneSelectorViewModel: Observable {
